requirements.txt
fastapi
uvicorn
sqlalchemy
databases
pydantic
pytest
httpx
python-dotenv
requests
